refactor(web): stop exporting everything from components/questions
