+++ web/src/components/questions/GenericQuestion.test.tsx
@@ -22,8 +22,8 @@
 import React from "react";
 import { screen } from "@testing-library/react";
 import { plainRender } from "~/test-utils";
-import { GenericQuestion } from "~/components/questions";
 import { Question } from "~/types/questions";
+import GenericQuestion from "~/components/questions/GenericQuestion";
 
 const question: Question = {
   id: 1,
