+++ web/src/components/questions/QuestionActions.test.tsx
@@ -22,8 +22,8 @@
 import React from "react";
 import { screen } from "@testing-library/react";
 import { installerRender } from "~/test-utils";
-import { QuestionActions } from "~/components/questions";
 import { Question } from "~/types/questions";
+import QuestionActions from "~/components/questions/QuestionActions";
 
 let defaultOption = "sure";
