+++ web/src/components/questions/index.ts
@@ -19,8 +19,4 @@
  * find current contact information at www.suse.com.
  */
 
-export { default as QuestionActions } from "./QuestionActions";
-export { default as GenericQuestion } from "./GenericQuestion";
-export { default as QuestionWithPassword } from "./QuestionWithPassword";
-export { default as LuksActivationQuestion } from "./LuksActivationQuestion";
 export { default as Questions } from "./Questions";
